pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas"
version = "0.1.0"
description = "Amortized Test-Time Learning from Adaptive Search"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "vllm",
    "openai",
    "triton",
    "pyyaml",
    "tqdm",
]

[tool.setuptools.packages.find]
include = ["benchmark*", "agent*", "experiment*", "utils*"]
